# Improve pure Go speed in low shard count