--- a/galois.go
+++ b/galois.go
@@ -8,6 +8,7 @@ package reedsolomon
 
 import (
 	"encoding/binary"
+	"sync"
 )
 
 const (
@@ -977,4 +978,28 @@ func sliceXorGo(in, out []byte, _ *options) {
 	for n, input := range in {
 		out[n] ^= input
 	}
+}
+
+// Combines 2 lookups into one.
+// 32MB in total.
+var mulTable16 *[256][65536]uint16
+var mulTable16Init sync.Once
+
+// getMulTable16 will return the 65536 entry table for the given column byte.
+func getMulTable16(c byte) *[65536]uint16 {
+	mulTable16Init.Do(func() {
+		mulTable16 = &[256][65536]uint16{}
+		// Generate two byte lookup table for multiplication
+		for i := 0; i < 256; i++ {
+			t0 := &mulTable[i]
+			t1 := &mulTable16[i]
+			for j := 0; j < 256; j++ {
+				for k := 0; k < 256; k++ {
+					dst := j*256 + k
+					t1[dst] = uint16(t0[j])<<8 | uint16(t0[k])
+				}
+			}
+		}
+	})
+	return &mulTable16[c]
 }