--- a/galois_noasm.go
+++ b/galois_noasm.go
@@ -12,9 +12,22 @@ func galMulSlice(c byte, in, out []byte, o *options) {
 		copy(out, in)
 		return
 	}
+	if !o.skip2B {
+		mt16 := getMulTable16(c)
+		for len(in) >= 8 {
+			store16(out, mt16[load16(in, 0)], 0)
+			store16(out, mt16[load16(in, 2)], 2)
+			store16(out, mt16[load16(in, 4)], 4)
+			store16(out, mt16[load16(in, 6)], 6)
+
+			in = in[8:]
+			out = out[8:]
+		}
+	}
+
 	mt := mulTable[c][:256]
 	for n, input := range in {
-		out[n] = mt[input]
+		store8(out, mt[input], n)
 	}
 }
 
@@ -24,9 +37,21 @@ func galMulSliceXor(c byte, in, out []byte, o *options) {
 		sliceXor(in, out, o)
 		return
 	}
+	if !o.skip2B {
+		mt16 := getMulTable16(c)
+		for len(in) >= 8 {
+			store16(out, load16(out, 0)^mt16[load16(in, 0)], 0)
+			store16(out, load16(out, 2)^mt16[load16(in, 2)], 2)
+			store16(out, load16(out, 4)^mt16[load16(in, 4)], 4)
+			store16(out, load16(out, 6)^mt16[load16(in, 6)], 6)
+
+			in = in[8:]
+			out = out[8:]
+		}
+	}
 	mt := mulTable[c][:256]
-	for n, input := range in {
-		out[n] ^= mt[input]
+	for n := range in {
+		store8(out, load8(out, n)^mt[load8(in, n)], n)
 	}
 }